requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
idna==3.3
pkg-resources==0.0.0
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
